Python soft IOC: Take copy of arrays being written to waveforms
Unfortunately it is too easy to make the mistake of writing a waveform
to a PV and then subsequently changing it.  This produces bugs which are
very hard to track down, and so this commit pays the cost of an extra
copy of the data to avoid this class of bug.

git-svn-id: svn+ssh://serv0002.cs.diamond.ac.uk/home/subversion/repos/controls/diamond/trunk/support/pythonSoftIoc@71355 e099a375-04f9-0310-9d5f-a741eaff62e1