Fix segmentation fault caused by 32-bit/64-bit mismatch
A pointer argument (to db_put_field) was losing its top 32-bits resulting
in a segmentation fault.  This was due to missing argtypes entries in the
imports definitions file.

This fault was not previously found for two reasons: firstly, the path in
error (calling .set() on an OUT record) is fairly rare in pracice; and
secondly, it turns out that a lot of data allocations are in the bottom
32-bits of address space!

git-svn-id: svn+ssh://serv0002.cs.diamond.ac.uk/home/subversion/repos/controls/diamond/trunk/support/pythonSoftIoc@152031 e099a375-04f9-0310-9d5f-a741eaff62e1