Fix regression introduced in release 2-12
This fixes a bug where records with names longer than 40 characters are not
correctly handled.

Commit c0420e4 added code to truncate all record string fields to 40
characters: this was done because it seems to be very unclear whether
EPICS string values are guaranteed to be null terminated.

Unfortunately, the same field access code is also used to retrieve the
record NAME field.

Reverting this part of the commit seems safe enough as we are simply
restoring the old long established behaviour.