Improve synchronisation for parent-child messages
This new model will have the parent process wait until IocInit() has
successfully completed before continuing with the test

The child process had need to do a queue.put() followed immediately by a
queue.get(). On most platforms the main process would interuupt by
pulling from the queue, but this was not guaranteed. To solve this using
queues would have required a second queue for signalling.

We work around this by using Pipes instead. These provide the
read-write semantics we want - the ability to wait for an explicit message
from the parent process - without needing a side-channel for more
signalling.

Hopefully this will fix the MacOS issues where it seemed caget was
returning values from a previous iteration of the tests, possibly
because the test process was running much faster than the child process
could keep up with.