Fix issue #78 by ensuring _read_value() returns true copy
It turns out that the value returned by record.read_val() can be a shared
reference to the value in the underlying EPICS record.  This means that
this value can change unexpectedly!

The fix is simply to apply the same logic as already done for string
records, and take a true copy instead.