Refactor tests to avoid most race conditions
These tests have passed 10 consecutive runs on CI with no failures.

There were race conditions regarding processing of record callbacks, so
we swap to a model of using a "DONE" record that, when processed,
signals that the IOC process has finished all of its callbacks.

Logging also added to some tests that had weird failures on CI, so if
they appear again we'll have more information.