Rework the handling of default values for in and out records
This commit addresses both github issues #81 and #67.

The arguments to the a and long record builder functions (`aIn`, `aOut`,
`longIn`, `longOut`) are brought more into line with the arguments as
used in the epics_device module.  This means that extra unnamed arguments
can be used to specify EGU and PREC (where appropriate), and for out
records the DRV limits take precedence.

As a side effect, the implementation of this is made closer to that in
https://github.com/dls-controls/epics_device/