Improve string representation of RecordWrapper
The returned string is the full DeviceName:RecordName.

Previously the string representation would be lost when
builder.LoadDatabase() is called, as the __builder instance is removed
and so __str__ returns None.

The __device is not removed at any point, and so can always be safely
used to fetch the full name.