Remove unnecessary delayed processing of value
Previously we did this as we wanted to spot the case where no value
was provided, either during record creation or by a set() call at some
later time, so that we could mark the record as being in an invalid
state.

Now that we store the severity and status alongside the value, we don't
need to handle this special case.